requirements.txt
kubernetes==28.1.0
PyYAML==6.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
streamlit==1.31.1

